refactor: moved flarect test case to one file
