+++ test/e2e-test/flareact.test.ts
@@ -48,7 +48,7 @@ describe('Create Flareact application', () => {
             }
         }
 
-        s3.deleteObjects({Bucket: bucketParams.Bucket, Delete: { Objects: keys}});
+        await s3.deleteObjects({Bucket: bucketParams.Bucket, Delete: { Objects: keys}}).promise();
     }
 
     before(async () => {
@@ -74,17 +74,55 @@ describe('Create Flareact application', () => {
         expect(stdout).to.be.equal('Completed.\n');
     });
 
-    it('Copy azion.json file ', async () => {
-        await copy(path.join(process.cwd(), 'test', 'config-files', 'azion-flareact.json'),path.join(template,'azion.json'))
-        const azionConfigFile = fs.existsSync(path.join(template,'azion.json'));
-        expect(azionConfigFile).to.be.true;
-    });
-
     it('Install packages', async () => {
         process.chdir(template);
         await execFile('npm install', {});
         const packageLock = fs.existsSync(path.join(template,'package-lock.json'));
         expect(packageLock).to.be.true;
+    });
+
+    it('Build the Flareact project without "azion.json" file and expect it fail', async () => {
+        const expectOnError =  "Couldn't read file 'azion.json' at the project's root directory. Because ENOENT: no such file or directory, open 'azion.json'\n"
+
+        function run(cmd: string) {
+            return new Promise((resolve) => {
+                execFile(cmd, (error: any, stdout: any, stderr: any) => {
+                    resolve({stdout,stderr, error})
+                })
+            })
+        }
+        const afaStdOutput = await run('azion-framework-adapter build');
+        const flareactOutputDir = fs.existsSync(path.join(template,'out/_flareact'));
+        expect((afaStdOutput as {stdout: string}).stdout).to.be.equal(expectOnError);
+        expect(flareactOutputDir).to.be.false;
+    });
+
+    it('Copy azion.json file ', async () => {
+        await copy(path.join(localOutput, 'test', 'config-files', 'azion-flareact-without-S3.json'),path.join(template,'azion.json'), { overwrite: true})
+        const azionConfigFile = fs.existsSync(path.join(template,'azion.json'));
+        expect(azionConfigFile).to.be.true;
+    });
+
+    it('Build the Flareact project with "azion.json" without S3 credentials', async () => {
+        const expectOnError =  "S3 credentials not set either in the configuration file or as environment variables.\n"
+
+        function run(cmd: string) {
+            return new Promise((resolve) => {
+                execFile(cmd, (error: any, stdout: any, stderr: any) => {
+                    resolve({stdout,stderr, error})
+                })
+            })
+        }
+        const afaStdOutput = await run('azion-framework-adapter build');
+        const flareactOutputDir = fs.existsSync(path.join(template,'out/_flareact'));
+        expect((afaStdOutput as {stdout: string}).stdout).to.be.equal(expectOnError);
+        expect(flareactOutputDir).to.be.false;
+    });
+
+    it('Copy azion.json file ', async () => {
+        await copy(path.join(localOutput, 'test', 'config-files', 'azion-flareact.json'),path.join(template,'azion.json'), { overwrite: true})
+        const azionConfigFile = fs.existsSync(path.join(template,'azion.json'));
+        expect(azionConfigFile).to.be.true;
     });
 
     it('Build the Flareact project', async () => {
@@ -98,17 +136,63 @@ describe('Create Flareact application', () => {
         expect(flareactOutputDir).to.be.true;
     });
 
-    it('Publish only asset of the flareact to S3', async () => {
-        removeAllKeys(bucketParams);
-        const { stdout } =await execFile('azion-framework-adapter publish -s');
-        const publishOutputMessage = `Loading asset manifest from ${template}/worker/manifest.json\n`;
-        expect(stdout).to.be.equal(publishOutputMessage);
-        const filesOnS3Bucket: any[] = [];
-        await getAllKeys(bucketParams, filesOnS3Bucket);
+    describe("Publish only assets", () => {
 
-        const manifest = fs.readFileSync(`${template}/worker/manifest.json`);
-        const manifestArray = Object.values(JSON.parse(manifest.toString()));
-        const filesOnS3BucketContainManifestFiles = manifestArray.every((each) =>  filesOnS3Bucket.includes(each) );
-        expect(filesOnS3BucketContainManifestFiles).to.be.true;
+        it('Publish only asset of the flareact to S3', async () => {
+            removeAllKeys(bucketParams);
+            const { stdout } =await execFile('azion-framework-adapter publish -s');
+            const publishOutputMessage = `Loading asset manifest from ${template}/worker/manifest.json\n`;
+            expect(stdout).to.be.equal(publishOutputMessage);
+            const filesOnS3Bucket: any[] = [];
+            await getAllKeys(bucketParams, filesOnS3Bucket);
+
+            const manifest = fs.readFileSync(`${template}/worker/manifest.json`);
+            const manifestArray = Object.values(JSON.parse(manifest.toString()));
+            const filesOnS3BucketContainManifestFiles = manifestArray.every((each) =>  filesOnS3Bucket.includes(each) );
+            expect(filesOnS3BucketContainManifestFiles).to.be.true;
+        });
+
+    });
+
+    describe("Publish only functions", async () => {
+
+        it('Copy azion.json file with wrong token ', async () => {
+            await copy(path.join(localOutput, 'test', 'config-files', 'azion-flareact-with-wrong-token.json'),path.join(template,'azion.json'), { overwrite: true})
+            const azionConfigFile = fs.existsSync(path.join(template,'azion.json'));
+            expect(azionConfigFile).to.be.true;
+        });
+
+        it('Try to publish only function with wrong token and expect to fail', async () => {
+            const expectOutput = 'Cannot save edge function to Azion: {"detail":"Invalid token"}\n';
+            function run(cmd: string) {
+                return new Promise((resolve) => {
+                    execFile(cmd, (error: any, stdout: any, stderr: any) => {
+                        resolve({stdout,stderr, error})
+                    })
+                })
+            }
+            const afaStdOutput = await run('azion-framework-adapter publish --only-function');
+            expect((afaStdOutput as {stdout: string}).stdout).to.be.equal(expectOutput);
+        })
+
+        it('Copy azion.json file with rigth token ', async () => {
+            await copy(path.join(localOutput, 'test', 'config-files', 'azion-flareact.json'),path.join(template,'azion.json'), { overwrite: true})
+            const azionConfigFile = fs.existsSync(path.join(template,'azion.json'));
+            expect(azionConfigFile).to.be.true;
+        });
+
+        it('Publish only function', async () => {
+            const expectOutput = 'Function id: 1\n';
+            function run(cmd: string) {
+                return new Promise((resolve) => {
+                    execFile(cmd, (error: any, stdout: any, stderr: any) => {
+                        resolve({stdout,stderr, error})
+                    })
+                })
+            }
+            const afaStdOutput = await run('azion-framework-adapter publish --only-function');
+            expect((afaStdOutput as {stdout: string}).stdout).to.be.equal(expectOutput);
+        })
+
     });
 })
